requirements.txt
South>=1.0
coverage
wheel
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
